GetImageUrl with resizing included
Images are hosted on images.contentful.com. For files on this host you
can attach the URI query parameters w and/or h to specify the desired
dimensions. The image will never be stretched, skewed or enlarged.
Instead it will be fit into the bounding box given by the w and h
parameters.
Additionaly, a q can be passed to define the JPEG compression quality
between 1 and 100 and the fm parameter can be used to change the format
to either "png" or "jpg".